requirements.txt
requests
pandas
streamlit
python-dotenv
